feat(web-ui): Orchestrators rename, dashboard landing, and admin/LLM-access consolidation
Operator-UI overhaul developed across the rename effort:

- Rename "Agents" -> "Orchestrators" throughout the operator UI (en + de),
  including nav, agent picker, and the operator agents page.
- Dashboard as the landing page: the root route now renders a dashboard
  (system-health strip, quick-access, role onboarding); chat moves to /chat.
- Business-case-first onboarding (businessCases.ts + dashboard component)
  replacing the technical bootstrap-profile flow.
- Admin consolidation: providers and subscription-CLI pages refactored into
  reusable panels (ProvidersPanel, SubscriptionClisPanel, LlmAccessTabs) with
  inline LLM key entry and in-page tab switching.
- Standard ("fallback") orchestrator protection: the operator agents API
  refuses to disable or delete it (409 fallback_protected); its seeded default
  name becomes "Standard Orchestrator".
- Store page copy + nav cleanup; chat turn-usage telemetry badge removed;
  header subtitle removed.
- CI: bundle the official Claude CLI into the published middleware image
  (INSTALL_SUBSCRIPTION_CLIS=true) so subscription-CLI agents work on self-host.

web-ui `next build` passes; middleware typecheck/lint/tests green.
